klients un admins nevsar izdzest pabeigto nomu

+++ resources/views/Noma.blade.php
@@ -239,7 +239,9 @@ document.addEventListener('DOMContentLoaded', function () {
                     @if(($item->PabeigsanasStatuss ?? null) !== 'Pabeigts')
                         <a href="/Noma/{{ $item->NomasID }}/edit" class="btn-action">Rediģēt</a>
                     @endif
-                    <a href="/Noma/{{ $item->NomasID }}/delete" onclick="return confirm('Vai tiešām vēlies dzēst šo ierakstu?');" class="btn-action">Dzēst</a>
+                    @if(Auth::check() && (($item->PabeigsanasStatuss ?? null) !== 'Pabeigts'))
+                        <a href="/Noma/{{ $item->NomasID }}/delete" onclick="return confirm('Vai tiešām vēlies dzēst šo ierakstu?');" class="btn-action">Dzēst</a>
+                    @endif
                 </div>
             </td>
           </tr>
